+++ src/DIPS.Xamarin.UI/Controls/Sheet/AlignmentOptions.cs
@@ -1,0 +1,18 @@
+﻿namespace DIPS.Xamarin.UI.Controls.Sheet
+{
+    /// <summary>
+    ///     The positions of the sheet when it appears.
+    /// </summary>
+    public enum AlignmentOptions
+    {
+        /// <summary>
+        ///     Positions the sheet at the bottom of the screen and will be draggable from bottom to top.
+        /// </summary>
+        Bottom = 0,
+
+        /// <summary>
+        ///     Positions the sheet at the top of the screen and will be draggable from top to bottom.
+        /// </summary>
+        Top
+    }
+}
