+++ src/DIPS.Xamarin.UI/DIPS.Xamarin.UI.csproj
@@ -42,6 +42,10 @@
       <Compile Update="Internal\Xaml\ToastView.xaml.cs">
         <DependentUpon>ToastView.xaml</DependentUpon>
       </Compile>
+      <Compile Update="Controls\Sheet\CheatSheet\CheatSheet.xaml.cs">
+        <DependentUpon>CheatSheet.xaml</DependentUpon>
+        <SubType>Code</SubType>
+      </Compile>
   </ItemGroup>
 
   <ItemGroup>
