+++ src/Samples/DIPS.Xamarin.UI.Samples/Controls/Sheet/SheetPage.xaml
@@ -32,8 +32,8 @@
                                 IsOpen="{Binding IsSheetOpen}"
                                 MaxPosition="{Binding MaxPosition}"
                                 MinPosition="{Binding MinPosition}"
-                                OnCloseCommand="{Binding OnCloseCommand}"
-                                OnCloseCommandParameter="Closed"
+                                CloseCommand="{Binding OnCloseCommand}"
+                                CloseCommandParameter="Closed"
                                 OnOpenCommand="{Binding OnOpenCommand}"
                                 OnOpenCommandParameter="Opened"
                                 Position="{Binding Position}"
